chore: configure java 25, fix failing tests, and update project settings
- Upgrade maven.compiler source/target to 25

- Upgrade junit-jupiter to 5.12.0 and maven-surefire-plugin to 3.5.5

- Ignore .sdkmanrc and GEMINI.md in .gitignore

- Add AGENTS.md (symlinked as GEMINI.md)

- Remove JVM OOM test to speed up test suite

- Fix logic errors in InsertionSort, ProductOfArrayExceptSelf, MaxNumberOfKSumPairs

- Fix incorrect test assertions in KthLargestElementArrayTest and NearestExitEntranceMazeTest

+++ src/main/java/leetcode75/twopointer/MaxNumberOfKSumPairs.java
@@ -13,7 +13,7 @@ public class MaxNumberOfKSumPairs {
     int rightPointer = nums.length - 1;
     int operationCounter = 0;
     while (leftPointer < rightPointer) {
-      int twoSum = nums[leftPointer] + nums[rightPointer];
+      long twoSum = (long) nums[leftPointer] + nums[rightPointer];
       if (twoSum > k) {
         rightPointer--;
       } else if (twoSum < k) {
